Don't include selected item in video tree breadcrumbs
End the breadcrumbs at the parent item instead of the selected item. For example, if you're in a directory and the selected item is a file in that directory, then the last element in the breadcrumbs is the directory name, not the file name. Presumably, the selected item will be highlighted in some way, so you don't need it to appear in the breadcrumbs also.